requirements.txt
bleach
blinker
cryptography
flask
flask-cors
flask-jwt-extended
flask-mail
flask-migrate
Flask-Security-Too
flask-sqlalchemy
mypy
phonenumbers
psycopg2-binary
pre-commit
python-dotenv
python-decouple
pyqrcode
SQLAlchemy-serializer
wikitextparser
wtforms
